Move some alignment tasks to new images.py file as methods on an Image class.
